rvi-sota-client: Update and fix rust-openssl
Update the version of RVI SOTA Client as well as
all crates on which it depends. Furthermore the
patch includes improvements for the deployment of
the systemd service and removes the old work
around for rust-openssl. Dbus has been moved from
RDEPENDS to DEPENDS to ensure that RVI SOTA client
will be built successfully.

Bug-AGL: SPEC-385

Change-Id: If276aac45b582d7d1ddb5b702fb4f40937d38e75
Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>