Move to AGL framework on top of systemd
This new version of the framework leverages systemd
features to launch its applications.

Some of the advantages are:
 - possible use of systemd features (namespace, cgroups, autostart,
   dependency resolution, socket activation, ...)
 - more feature are let open to integrator's design

Some of the drawbacks are:
 - not more possible to launch an other instance of an application
   already launched
 - pause/resume is no more available by the framework
 - the remote mode is to be redefined

This commit integrates the first version of the framework on
top of systemd. More work is to come.

None of the current drawback is definitive.

Bug-AGL: SPEC-138, SPEC-425, SPEC-426, SPEC-427

Change-Id: Idfb98761c0db23562bb783bed1b03aeb956fc587
Signed-off-by: José Bollo <jose.bollo@iot.bzh>